# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Dollar CARE'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
...
  /api/v1/doctor/manage-clinic/presentSheduleDate:
    post:
      summary: 'Present shedule date'
      operationId: presentSheduleDate
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n\"status\": true,\n \"statusCode\": 200,\n \"message\": \"Clinic list\",\n \"data\": {\n         {\n             \"id\": 51,\n             \"slot_start_time\": \"2:00 PM\",\n             \"slot_end_time\": \"5:00 PM\",\n             \"slot_date\": \"2023-09-21\"\n         },\n         {\n             \"id\": 50,\n             \"slot_start_time\": \"1:00 PM\",\n             \"slot_end_time\": \"3:00 PM\",\n             \"slot_date\": \"2023-09-19\"\n         },\n }\n}"
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: false
                  statusCode: 201
                  error: 'No clinic found'
                properties:
                  status:
                    type: boolean
                    example: false
                  statusCode:
                    type: integer
                    example: 201
                  error:
                    type: string
                    example: 'No clinic found'
      tags:
        - 'Manage Clinic'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: numeric
                  description: ''
                  example: quibusdam
              required:
                - id
  /api/v1/doctor/manage-clinic/deletePresentSheduleDate:
    post:
      summary: 'Delete Present shedule date'
      operationId: deletePresentSheduleDate
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n\"status\": true,\n \"statusCode\": 200,\n\"message\": \"Slot deleted successfully\",\n}\n}"
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: false
                  statusCode: 201
                  error: 'No clinic found'
                properties:
                  status:
                    type: boolean
                    example: false
                  statusCode:
                    type: integer
                    example: 201
                  error:
                    type: string
                    example: 'No clinic found'
      tags:
        - 'Manage Clinic'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: numeric
                  description: ''
                  example: rerum
              required:
                - id
...